basic refactor complete. Needs review, docs, and some new tests.
