+++ primers/primers_test.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/TimothyStiles/poly/transformations"
+	"github.com/TimothyStiles/poly/transform/reverse"
 )
 
 func ExampleMarmurDoty() {
@@ -52,7 +52,7 @@ func TestSantaLucia(t *testing.T) {
 func TestSantaLuciaReverseComplement(t *testing.T) {
 	testSeq := "ACGTAGATCTACGT" //"GTAAAACGACGGCCAGT" // M13 fwd
 
-	testReverseComplement := transformations.ReverseComplement(testSeq)
+	testReverseComplement := reverse.ReverseComplement(testSeq)
 	if testSeq != testReverseComplement {
 		t.Errorf("Input is not a reverse complement of it's. Got %q instead of %q", testSeq, testReverseComplement)
 	}
@@ -121,7 +121,7 @@ func TestCreateBarcode(t *testing.T) {
 		t.Errorf("TestUniqueSequence string should return CTCTCGGTCGCTCCGTCCCG. Got:\n%s", output)
 	}
 
-	barcodes = CreateBarcodesWithBannedSequences(20, 4, []string{transformations.ReverseComplement("GGCCGCGCCCC")}, []func(string) bool{})
+	barcodes = CreateBarcodesWithBannedSequences(20, 4, []string{reverse.ReverseComplement("GGCCGCGCCCC")}, []func(string) bool{})
 	output = barcodes[len(barcodes)-1]
 	if output != "CTCTCGGTCGCTCCGTCCCG" {
 		t.Errorf("TestUniqueSequence string should return CTCTCGGTCGCTCCGTCCCG. Got:\n%s", output)
