+++ transform/codon/codon_test.go
@@ -1,46 +1,15 @@
-package transformations
+package codon
 
 import (
 	"fmt"
 	"io/ioutil"
-	"log"
 	"os"
-	"sort"
 	"strings"
 	"testing"
 
 	"github.com/TimothyStiles/poly/io/genbank"
 	"github.com/google/go-cmp/cmp"
 )
-
-func TestIUPAC(t *testing.T) {
-	testSeq := "ATN"
-	testVariants := []string{"ATG", "ATA", "ATT", "ATC"}
-	testVariantsIUPAC, err := AllVariantsIUPAC(testSeq)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	sort.Strings(testVariants)
-	sort.Strings(testVariantsIUPAC)
-
-	for index := range testVariants {
-		if testVariants[index] != testVariantsIUPAC[index] {
-			t.Errorf("IUPAC variant has changed on test 'allIUPAC('ATN')'. Got this:\n%s instead of \n%s", testVariantsIUPAC, testVariants)
-		}
-	}
-}
-
-func ExampleAllVariantsIUPAC() {
-	// AllVariantsIUPAC takes a string as input
-	// and returns all iupac variants as output
-	mendelIUPAC := "ATGGARAAYGAYGARCTN"
-	// ambiguous IUPAC codes for most of the sequences that code for the protein MENDEL
-	mendelIUPACvariants, _ := AllVariantsIUPAC(mendelIUPAC)
-	fmt.Println(mendelIUPACvariants)
-	// Output: [ATGGAGAATGATGAGCTG ATGGAGAATGATGAGCTA ATGGAGAATGATGAGCTT ATGGAGAATGATGAGCTC ATGGAGAATGATGAACTG ATGGAGAATGATGAACTA ATGGAGAATGATGAACTT ATGGAGAATGATGAACTC ATGGAGAATGACGAGCTG ATGGAGAATGACGAGCTA ATGGAGAATGACGAGCTT ATGGAGAATGACGAGCTC ATGGAGAATGACGAACTG ATGGAGAATGACGAACTA ATGGAGAATGACGAACTT ATGGAGAATGACGAACTC ATGGAGAACGATGAGCTG ATGGAGAACGATGAGCTA ATGGAGAACGATGAGCTT ATGGAGAACGATGAGCTC ATGGAGAACGATGAACTG ATGGAGAACGATGAACTA ATGGAGAACGATGAACTT ATGGAGAACGATGAACTC ATGGAGAACGACGAGCTG ATGGAGAACGACGAGCTA ATGGAGAACGACGAGCTT ATGGAGAACGACGAGCTC ATGGAGAACGACGAACTG ATGGAGAACGACGAACTA ATGGAGAACGACGAACTT ATGGAGAACGACGAACTC ATGGAAAATGATGAGCTG ATGGAAAATGATGAGCTA ATGGAAAATGATGAGCTT ATGGAAAATGATGAGCTC ATGGAAAATGATGAACTG ATGGAAAATGATGAACTA ATGGAAAATGATGAACTT ATGGAAAATGATGAACTC ATGGAAAATGACGAGCTG ATGGAAAATGACGAGCTA ATGGAAAATGACGAGCTT ATGGAAAATGACGAGCTC ATGGAAAATGACGAACTG ATGGAAAATGACGAACTA ATGGAAAATGACGAACTT ATGGAAAATGACGAACTC ATGGAAAACGATGAGCTG ATGGAAAACGATGAGCTA ATGGAAAACGATGAGCTT ATGGAAAACGATGAGCTC ATGGAAAACGATGAACTG ATGGAAAACGATGAACTA ATGGAAAACGATGAACTT ATGGAAAACGATGAACTC ATGGAAAACGACGAGCTG ATGGAAAACGACGAGCTA ATGGAAAACGACGAGCTT ATGGAAAACGACGAGCTC ATGGAAAACGACGAACTG ATGGAAAACGACGAACTA ATGGAAAACGACGAACTT ATGGAAAACGACGAACTC]
-
-}
 
 func ExampleTranslate() {
 
@@ -101,10 +70,11 @@ func ExampleOptimize() {
 
 	gfpTranslation := "MASKGEELFTGVVPILVELDGDVNGHKFSVSGEGEGDATYGKLTLKFICTTGKLPVPWPTLVTTFSYGVQCFSRYPDHMKRHDFFKSAMPEGYVQERTISFKDDGNYKTRAEVKFEGDTLVNRIELKGIDFKEDGNILGHKLEYNYNSHNVYITADKQKNGIKANFKIRHNIEDGSVQLADHYQQNTPIGDGPVLLPDNHYLSTQSALSKDPNEKRDHMVLLEFVTAAGITHGMDELYK*"
 
-	sequence := genbank.Read("data/puc19.gbk")
+	sequence := genbank.Read("../../data/puc19.gbk")
 	codonTable := GetCodonTable(11)
+	codingRegions := GetCodingRegions(sequence)
 
-	optimizationTable := sequence.GetOptimizationTable(codonTable)
+	optimizationTable := codonTable.OptimizeTable(codingRegions)
 
 	optimizedSequence, _ := Optimize(gfpTranslation, optimizationTable)
 	optimizedSequenceTranslation, _ := Translate(optimizedSequence, optimizationTable)
@@ -116,10 +86,11 @@ func ExampleOptimize() {
 func TestOptimize(t *testing.T) {
 	gfpTranslation := "MASKGEELFTGVVPILVELDGDVNGHKFSVSGEGEGDATYGKLTLKFICTTGKLPVPWPTLVTTFSYGVQCFSRYPDHMKRHDFFKSAMPEGYVQERTISFKDDGNYKTRAEVKFEGDTLVNRIELKGIDFKEDGNILGHKLEYNYNSHNVYITADKQKNGIKANFKIRHNIEDGSVQLADHYQQNTPIGDGPVLLPDNHYLSTQSALSKDPNEKRDHMVLLEFVTAAGITHGMDELYK*"
 
-	sequence := ReadGbk("data/puc19.gbk")
+	sequence := genbank.Read("../../data/puc19.gbk")
 	codonTable := GetCodonTable(11)
+	codingRegions := GetCodingRegions(sequence)
 
-	optimizationTable := sequence.GetOptimizationTable(codonTable)
+	optimizationTable := codonTable.OptimizeTable(codingRegions)
 
 	optimizedSequence, _ := Optimize(gfpTranslation, optimizationTable)
 	optimizedSequenceTranslation, _ := Translate(optimizedSequence, optimizationTable)
@@ -200,14 +171,14 @@ JSON related tests begin here.
 ******************************************************************************/
 
 func ExampleReadCodonJSON() {
-	codontable := ReadCodonJSON("data/bsub_codon_test.json")
+	codontable := ReadCodonJSON("../../data/bsub_codon_test.json")
 
 	fmt.Println(codontable.AminoAcids[0].Codons[0].Weight)
 	//output: 28327
 }
 
 func ExampleParseCodonJSON() {
-	file, _ := ioutil.ReadFile("data/bsub_codon_test.json")
+	file, _ := ioutil.ReadFile("../../data/bsub_codon_test.json")
 	codontable := ParseCodonJSON(file)
 
 	fmt.Println(codontable.AminoAcids[0].Codons[0].Weight)
@@ -215,24 +186,24 @@ func ExampleParseCodonJSON() {
 }
 
 func ExampleWriteCodonJSON() {
-	codontable := ReadCodonJSON("data/bsub_codon_test.json")
-	WriteCodonJSON(codontable, "data/codon_test.json")
-	testCodonTable := ReadCodonJSON("data/codon_test.json")
+	codontable := ReadCodonJSON("../../data/bsub_codon_test.json")
+	WriteCodonJSON(codontable, "../../data/codon_test.json")
+	testCodonTable := ReadCodonJSON("../../data/codon_test.json")
 
 	// cleaning up test data
-	os.Remove("data/codon_test.json")
+	os.Remove("../../data/codon_test.json")
 
 	fmt.Println(testCodonTable.AminoAcids[0].Codons[0].Weight)
 	//output: 28327
 }
 
 func TestWriteCodonJSON(t *testing.T) {
-	testCodonTable := ReadCodonJSON("data/bsub_codon_test.json")
-	WriteCodonJSON(testCodonTable, "data/codon_test1.json")
-	readTestCodonTable := ReadCodonJSON("data/codon_test1.json")
+	testCodonTable := ReadCodonJSON("../../data/bsub_codon_test.json")
+	WriteCodonJSON(testCodonTable, "../../data/codon_test1.json")
+	readTestCodonTable := ReadCodonJSON("../../data/codon_test1.json")
 
 	// cleaning up test data
-	os.Remove("data/codon_test1.json")
+	os.Remove("../../data/codon_test1.json")
 
 	if diff := cmp.Diff(testCodonTable, readTestCodonTable); diff != "" {
 		t.Errorf(" mismatch (-want +got):\n%s", diff)
@@ -247,13 +218,15 @@ Codon Compromise + Add related tests begin here.
 ******************************************************************************/
 
 func ExampleCompromiseCodonTable() {
-	sequence := ReadGbk("data/puc19.gbk")
+	sequence := genbank.Read("../../data/puc19.gbk")
 	codonTable := GetCodonTable(11)
-	optimizationTable := sequence.GetOptimizationTable(codonTable)
+	codingRegions := GetCodingRegions(sequence)
+	optimizationTable := codonTable.OptimizeTable(codingRegions)
 
-	sequence2 := ReadGbk("data/phix174.gb")
+	sequence2 := genbank.Read("../../data/phix174.gb")
 	codonTable2 := GetCodonTable(11)
-	optimizationTable2 := sequence2.GetOptimizationTable(codonTable2)
+	codingRegions2 := GetCodingRegions(sequence2)
+	optimizationTable2 := codonTable2.OptimizeTable(codingRegions2)
 
 	finalTable, _ := CompromiseCodonTable(optimizationTable, optimizationTable2, 0.1)
 	for _, aa := range finalTable.AminoAcids {
@@ -267,13 +240,15 @@ func ExampleCompromiseCodonTable() {
 }
 
 func TestCompromiseCodonTable(t *testing.T) {
-	sequence := ReadGbk("data/puc19.gbk")
+	sequence := genbank.Read("../../data/puc19.gbk")
 	codonTable := GetCodonTable(11)
-	optimizationTable := sequence.GetOptimizationTable(codonTable)
+	codingRegions := GetCodingRegions(sequence)
+	optimizationTable := codonTable.OptimizeTable(codingRegions)
 
-	sequence2 := ReadGbk("data/phix174.gb")
+	sequence2 := genbank.Read("../../data/phix174.gb")
 	codonTable2 := GetCodonTable(11)
-	optimizationTable2 := sequence2.GetOptimizationTable(codonTable2)
+	codingRegions2 := GetCodingRegions(sequence2)
+	optimizationTable2 := codonTable2.OptimizeTable(codingRegions2)
 
 	_, err := CompromiseCodonTable(optimizationTable, optimizationTable2, -1.0) // Fails too low
 	if err == nil {
@@ -286,13 +261,15 @@ func TestCompromiseCodonTable(t *testing.T) {
 }
 
 func ExampleAddCodonTable() {
-	sequence := ReadGbk("data/puc19.gbk")
+	sequence := genbank.Read("../../data/puc19.gbk")
 	codonTable := GetCodonTable(11)
-	optimizationTable := sequence.GetOptimizationTable(codonTable)
+	codingRegions := GetCodingRegions(sequence)
+	optimizationTable := codonTable.OptimizeTable(codingRegions)
 
-	sequence2 := ReadGbk("data/phix174.gb")
+	sequence2 := genbank.Read("../../data/phix174.gb")
 	codonTable2 := GetCodonTable(11)
-	optimizationTable2 := sequence2.GetOptimizationTable(codonTable2)
+	codingRegions2 := GetCodingRegions(sequence2)
+	optimizationTable2 := codonTable2.OptimizeTable(codingRegions2)
 
 	finalTable := AddCodonTable(optimizationTable, optimizationTable2)
 	for _, aa := range finalTable.AminoAcids {
